refactored OcAvatarGroup to OcAvatars
