+++ src/components/avatars/OcAvatarLink.vue
@@ -1,5 +1,5 @@
 <template>
-  <oc-avatar-item background="#C6CFD9" icon="link" :name="name" :accessible-label="accessibleLabel"></oc-avatar-item>
+  <oc-avatar-item icon="link" :name="name" :accessible-label="accessibleLabel"></oc-avatar-item>
 </template>
 
 <script>
@@ -26,7 +26,7 @@ export default {
       required: false,
       default: "",
     },
-  }
+  },
 }
 </script>
